# conv_direct!(): The performance fix