--- a/src/impl/depthwiseconv_direct.jl
+++ b/src/impl/depthwiseconv_direct.jl
@@ -18,7 +18,7 @@ channels in `x` is the last, not the second-to-last, as in a normal dense convol
 
 See the docstring for `conv_direct!()` for more on the optional parameters.
 """
-function depthwiseconv_direct!(
+function depthwiseconv_direct_old!(
                 y::AbstractArray{yT,5}, x::AbstractArray{xT,5},
                 w::AbstractArray{wT,5}, cdims::DepthwiseConvDims;
                 alpha::yT = yT(1), beta::yT = yT(0)) where {yT, xT, wT}
@@ -80,6 +80,108 @@ function depthwiseconv_direct!(
         y[w_idx, h_idx, d_idx, c_out, batch] = alpha*convert(yT, dotprod) + prev_yval
     end
     
+    return y
+end
+
+function depthwiseconv_direct!(y::AbstractArray{yT,5}, x::AbstractArray{xT,5},
+                      w::AbstractArray{wT,5}, cdims::DepthwiseConvDims;
+                      alpha::yT = yT(1), beta = false) where {yT, xT, wT}
+    check_dims(size(x), size(w), size(y), cdims)
+
+    width, height, depth = input_size(cdims)
+    kernel_w, kernel_h, kernel_d = kernel_size(cdims)
+    out_c = channels_out(cdims)
+    pad_w_lo, pad_w_hi, pad_h_lo, pad_h_hi, pad_d_lo, pad_d_hi = padding(cdims)
+    dil_w, dil_h, dil_d = dilation(cdims)
+    stride_w, stride_h, stride_d = stride(cdims)
+    out_width, out_height, out_depth = output_size(cdims)
+    
+    # Create a method that, at compile-time, determines how we're going to index into `w`
+    kproj(k, M, cdims::DepthwiseConvDims{N,K,C_mult,C_in,S,P,D,true}) where {N, K, C_mult, C_in, S, P, D} = k
+    kproj(k, M, cdims::DepthwiseConvDims{N,K,C_mult,C_in,S,P,D,false}) where {N, K, C_mult, C_in, S, P, D} = M - k + 1
+    
+    # A helper function to project from output (w, h) to input (input_w, input_h)
+    project(idx, stride, pad) = (idx - 1)*stride - pad + 1
+    
+    # Use `calc_padding_regions` to determine where we do or don't need to worry about padding
+    padded_regions, central_region = calc_padding_regions(cdims)
+
+    # Start with the central region
+    w_region, h_region, d_region = central_region
+    @inbounds for batch in 1:size(x)[end],
+        c_mult in 1:channel_multiplier(cdims),
+        c_in in 1:channels_in(cdims),
+        d_idx in d_region,
+        h_idx in h_region,
+        w_idx in w_region
+
+        # Since we're in the central region, we don't need to worry about clamping
+        dotprod = yT(0)
+        c_out = (c_in - 1)*channel_multiplier(cdims) + c_mult
+        for kd in 1:kernel_d,
+            kh in 1:kernel_h,
+            kw in 1:kernel_w
+
+            # Hoist me, you coward.
+            x_d = project(d_idx, stride_d, pad_d_lo) + (kd - 1)*dil_d
+            x_h = project(h_idx, stride_h, pad_h_lo) + (kh - 1)*dil_h
+            x_w = project(w_idx, stride_w, pad_w_lo) + (kw - 1)*dil_w
+
+            x_val = x[x_w, x_h, x_d, c_in, batch]
+            w_val = w[kproj(kw, kernel_w, cdims),
+                      kproj(kh, kernel_h, cdims),
+                      kproj(kd, kernel_d, cdims),
+                      c_mult, c_in]
+            dotprod = muladd(x_val, w_val, dotprod)
+        end
+        y[w_idx, h_idx, d_idx, c_out, batch] = alpha*dotprod + beta*y[w_idx, h_idx, d_idx, c_out, batch]
+    end
+
+    # Next, do potentially-padded regions:
+    @inbounds for (w_region, h_region, d_region) in padded_regions,
+        batch in 1:size(x)[end],
+        c_mult in 1:channel_multiplier(cdims),
+        c_in in 1:channels_in(cdims),
+        d_idx in d_region,
+        h_idx in h_region,
+        w_idx in w_region
+
+        # Probe for out-of-bounds accesses on `x` and `continue` if we hit one
+        dotprod = yT(0)
+        c_out = (c_in - 1)*channel_multiplier(cdims) + c_mult
+        for c_in in 1:channels_in(cdims),
+            kd in 1:kernel_d
+
+            x_d = project(d_idx, stride_d, pad_d_lo) + (kd - 1)*dil_d
+            if x_d <= 0 || x_d > depth
+                continue
+            end
+
+            for kh in 1:kernel_h
+                x_h = project(h_idx, stride_h, pad_h_lo) + (kh - 1)*dil_h
+                if x_h <= 0 || x_h > height
+                    continue
+                end
+
+                for kw in 1:kernel_w
+                    x_w = project(w_idx, stride_w, pad_w_lo) + (kw - 1)*dil_w
+                    if x_w <= 0 || x_w > width
+                        continue
+                    end
+
+                    x_val = x[x_w, x_h, x_d, c_in, batch]
+                    w_val = w[kproj(kw, kernel_w, cdims),
+                              kproj(kh, kernel_h, cdims),
+                              kproj(kd, kernel_d, cdims),
+                              c_mult, c_in]
+                    dotprod = muladd(x_val, w_val, dotprod)
+                end
+            end
+        end
+
+        y[w_idx, h_idx, d_idx, c_out, batch] = alpha*dotprod + beta*y[w_idx, h_idx, d_idx, c_out, batch]
+    end
+
     return y
 end